+++ source/savefile.cpp
@@ -247,7 +247,7 @@ namespace savefile {
         }
     }
 
-    void check_players(ifstream &dream_file, u32 account_offset, bool *g_players) {
+    void check_players(ifstream &dream_file, u64 account_offset, bool *g_players) {
         for (u8 player = 0; player < 8; player++) {
             u64 offset = player * 0x48;
             u128 account_uid = 0;
